requirements.txt
flask==3.0.0
dnspython==2.4.2
rich==13.7.0
